data-integration/from-github-to-firestore/requirements.txt
requests
markdown2
grpcio
google-cloud-core
google-cloud-firestore